+++ src/r0/typing/create_typing_event.rs
@@ -1,5 +1,6 @@
 //! [PUT /_matrix/client/r0/rooms/{roomId}/typing/{userId}](https://matrix.org/docs/spec/client_server/r0.4.0.html#put-matrix-client-r0-rooms-roomid-typing-userid)
 
+use js_int::UInt;
 use ruma_api_macros::ruma_api;
 use ruma_identifiers::{RoomId, UserId};
 use serde::{Deserialize, Serialize};
@@ -20,7 +21,7 @@ ruma_api! {
         pub room_id: RoomId,
         /// The length of time in milliseconds to mark this user as typing.
         #[serde(skip_serializing_if = "Option::is_none")]
-        pub timeout: Option<u64>,
+        pub timeout: Option<UInt>,
         /// Whether the user is typing or not. If `false`, the `timeout` key can be omitted.
         pub typing: bool,
         /// The user who has started to type.
